sam2/server/client/requirements_client.txt
# HTTP requests
requests
aiohttp

# Image processing
Pillow
numpy
opencv-python

# Configuration and data
pyyaml

# Visualization (optional)
matplotlib

# Additional utilities
typing-extensions 